+++ views/admin/addressbook/delete.html
@@ -1,4 +1,4 @@
-{% extends 'admin/layout.html' %}
+{% extends 'layout.html' %}
 {% block content %}
 <header class="jumbotron subhead" id="overview">
     <h1><i class="glyph2x-adress-book"></i>Delete address book</h1>
