+++ views/admin/dashboard.html
@@ -1,4 +1,4 @@
-{% extends 'admin/layout.html' %}
+{% extends 'layout.html' %}
 {% block content %}
     {% set url = 'http://baikal-server.com' %}
     <style>
