+++ views/layout.html
@@ -20,7 +20,7 @@
 </head>
 <body>
 {% block navbar %}
-    {% embed "admin/navbar.html" %}
+    {% embed "navbar.html" %}
     {% endembed %}
 {% endblock %}
