+++ views/login.html
@@ -1,11 +1,11 @@
-{% extends 'admin/layout.html' %}
+{% extends 'layout.html' %}
 {% block content %}
     <header class="jumbotron subhead" id="overview">
         <h1><i class="glyph2x-lock"></i>Authentication</h1>
         <p class="lead">Please authenticate to access Baikal Web Admin.</p>
     </header>
     {# {{ msg }} #}
-    <form class="form-horizontal" action="{{ path("admin_login_post") }}" method="post">
+    <form class="form-horizontal" action="{{ path("login") }}" method="post">
         <fieldset>
 
 			<div class="control-group">
