+++ views/navbar.html
@@ -9,10 +9,12 @@
             <a class="brand" href="{{ path('home') }}"><img style="vertical-align: text-top; line-height: 20px;" src="{{ assetPath }}core/BaikalAdmin/Templates/Page/baikal-text-20.png" /> Web Admin</a>
             <div class="nav-collapse">
                 <ul class="nav">
+                    {% if authenticated %}
                     <li class=""> <a href="{{ path('admin_dashboard') }}">Dashboard</a></li>
                     <li class=""> <a href="{{ path('admin_user_index') }}">Users and resources</a></li>
                     <li class=""> <a href="{{ path('admin_settings') }}">Settings</a></li>
-                    <li> <a href="{{ path('admin_logout') }}"><i class="icon-eject icon-white"></i> Logout</a></li>
+                    <li> <a href="{{ path('logout') }}"><i class="icon-eject icon-white"></i> Logout</a></li>
+                    {% endif %}
                 </ul>
             </div>
         </div>
